+++ src/tc/tc.h
@@ -1,2 +1,2 @@
 #include "common.h"
-void TCSolver(int m, int nnz, int *row_offsets, int *column_indices, int *degree, size_t *total);
+void TCSolver(int m, int nnz, int *row_offsets, int *column_indices, int *degree, int *total);
